chore: Update npm dependencies and add uuid and @types/uuid
The code changes involve updating the npm dependencies in the package.json and package-lock.json files. The uuid package has been added with version 9.0.1 and the @types/uuid package has been added with version 9.0.8. These packages are necessary for generating and working with universally unique identifiers (UUIDs) in the application.

Note: The commit message follows the conventional format of starting with a type (chore) followed by a brief description of the changes.