chore: Update Select component changeOption prop to be optional
The code changes involve updating the Select component in the Select.tsx file. The changeOption prop has been changed to be optional by adding a question mark (?) after its name. This change allows for flexibility in using the Select component, as the changeOption prop can now be provided or omitted based on the specific use case.

Note: The commit message follows the conventional format of starting with a type (chore) followed by a brief description of the changes.